+++ reframe/core/schedulers/slurm.py
@@ -249,34 +249,36 @@ class SlurmJobScheduler(sched.JobScheduler):
         return list(filter(None, preamble))
 
     def submit(self, job):
-        cmd = f'sbatch {job.script_filename}'
-        intervals = itertools.cycle([1, 2, 3])
-        while True:
-            try:
-                completed = _run_strict(cmd, timeout=self._submit_timeout)
-                break
-            except SpawnedProcessError as e:
-                error_match = re.search(
-                    rf'({"|".join(self._resubmit_on_errors)})', e.stderr
+        if not job.dry_run_mode:
+            cmd = f'sbatch {job.script_filename}'
+            intervals = itertools.cycle([1, 2, 3])
+            while True:
+                try:
+                    completed = _run_strict(cmd, timeout=self._submit_timeout)
+                    break
+                except SpawnedProcessError as e:
+                    error_match = re.search(
+                        rf'({"|".join(self._resubmit_on_errors)})', e.stderr
+                    )
+                    if not self._resubmit_on_errors or not error_match:
+                        raise
+
+                    t = next(intervals)
+                    self.log(
+                        f'encountered a job submission error: '
+                        f'{error_match.group(1)}: will resubmit after {t}s'
+                    )
+                    time.sleep(t)
+
+            jobid_match = re.search(r'Submitted batch job (?P<jobid>\d+)',
+                                    completed.stdout)
+            if not jobid_match:
+                raise JobSchedulerError(
+                    'could not retrieve the job id of the submitted job'
                 )
-                if not self._resubmit_on_errors or not error_match:
-                    raise
 
-                t = next(intervals)
-                self.log(
-                    f'encountered a job submission error: '
-                    f'{error_match.group(1)}: will resubmit after {t}s'
-                )
-                time.sleep(t)
+            job._jobid = jobid_match.group('jobid')
 
-        jobid_match = re.search(r'Submitted batch job (?P<jobid>\d+)',
-                                completed.stdout)
-        if not jobid_match:
-            raise JobSchedulerError(
-                'could not retrieve the job id of the submitted job'
-            )
-
-        job._jobid = jobid_match.group('jobid')
         job._submit_time = time.time()
 
     def allnodes(self):
@@ -402,6 +404,12 @@ class SlurmJobScheduler(sched.JobScheduler):
 
     def poll(self, *jobs):
         '''Update the status of the jobs.'''
+
+        if jobs:
+            for job in jobs:
+                if job.dry_run_mode:
+                    job._state = 'COMPLETED'
+                    job._completed = True
 
         if jobs:
             # Filter out non-jobs
@@ -551,7 +559,9 @@ class SlurmJobScheduler(sched.JobScheduler):
             self._merge_files(job)
 
     def cancel(self, job):
-        _run_strict(f'scancel {job.jobid}', timeout=self._submit_timeout)
+        if not job.dry_run_mode:
+            _run_strict(f'scancel {job.jobid}', timeout=self._submit_timeout)
+
         job._is_cancelling = True
 
     def finished(self, job):
@@ -568,6 +578,12 @@ class SqueueJobScheduler(SlurmJobScheduler):
     SQUEUE_DELAY = 2
 
     def poll(self, *jobs):
+        if jobs:
+            for job in jobs:
+                if job.dry_run_mode:
+                    job._state = 'COMPLETED'
+                    job._completed = True
+
         if jobs:
             # Filter out non-jobs
             jobs = [job for job in jobs if job is not None]
